Move all db models into db dir
