+++ client/src/shared/models/db/FlattenedEnrollment.ts
@@ -1,4 +1,4 @@
-import { EnrollmentReport } from '.';
+import { EnrollmentReport } from '..';
 import { Moment } from 'moment';
 
 export interface FlattenedEnrollment {
