+++ client/src/shared/models/db/Funding.ts
