+++ client/src/shared/models/db/FundingTimeSplit.ts
