+++ client/src/shared/models/db/ReportingPeriod.ts
@@ -1,4 +1,4 @@
-import { FundingSource } from '.';
+import { FundingSource } from '../';
 import { Moment } from 'moment';
 
 export interface ReportingPeriod {
