+++ client/src/shared/models/db/index.ts
@@ -1,0 +1,16 @@
+export * from './Child';
+export * from './Community';
+export * from './Enrollment';
+export * from './EnrollmentReport';
+export * from './Family';
+export * from './FlattenedEnrollment';
+export * from './Funding';
+export * from './FundingSpace';
+export * from './FundingTimeSplit';
+export * from './IncomeDetermination';
+export * from './Organization';
+export * from './OrganizationPermission';
+export * from './ReportingPeriod';
+export * from './Site';
+export * from './SitePermission';
+export * from './User';
