+++ client/src/shared/models/index.ts
@@ -1,25 +1,11 @@
+export * from './db';
+
 export * from './AgeGroup';
-export * from './Child';
-export * from './Community';
 export * from './ColumnMetadata';
-export * from './Enrollment';
-export * from './EnrollmentReport';
 export * from './ExitReason';
-export * from './Family';
-export * from './FlattenedEnrollment';
-export * from './Funding';
 export * from './FundingSource';
-export * from './FundingSpace';
 export * from './FundingSourceTime';
 export * from './FundingTime';
-export * from './FundingTimeSplit';
 export * from './Gender';
-export * from './IncomeDetermination';
-export * from './Organization';
-export * from './OrganizationPermission';
 export * from './Region';
-export * from './ReportingPeriod';
-export * from './Site';
-export * from './SitePermission';
 export * from './SpecialEducationServicesType';
-export * from './User';
